feat: support uint64 token bucket rates

+++ include/click/bigint.hh
@@ -385,7 +385,7 @@ class Bigint { public:
 };
 
 /** @brief Typical Bigint usage with uint32_t limb_type. */
-typedef Bigint<uint32_t> bigint;
+typedef Bigint<uint64_t> bigint;
 
 CLICK_ENDDECLS
 #endif
